TSDK-802 Fix unit tests
